requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.1.0
av==15.1.0
bert-score==0.3.13
bitsandbytes==0.48.1
blake3==1.0.8
cachetools==6.2.1
cbor2==5.7.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
compressed-tensors==0.11.0
contourpy==1.3.0
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
interegular==0.3.3
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
llguidance==0.7.30
llvmlite==0.43.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mistral_common==1.8.5
modelscope==1.28.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.2.1
ninja==1.13.0
nltk==3.9.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==2.3.0
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
outlines_core==0.2.11
packaging==25.0
pandas==2.3.1
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.3.2
protobuf==6.33.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.2
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qwen-omni-utils==0.0.8
qwen-vl-utils==0.0.14
ray==2.50.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rouge==1.0.1
rpds-py==0.27.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.1
sentry-sdk==2.42.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.1
triton==3.4.0
trl==0.24.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.1
websockets==15.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
